Add vuex to project
Needed a way to deal with router-view and passing props.
One way was with event-bus,
other with route props(which is not good way for events),
and finally vuex.
Took vuex since it's used in bigger projects,
and it provides simple, yet safe use in terms of single source of truth.